UnorderedForwarder.sol

+++ contracts/marketplace/base/MarketplaceBase.sol
@@ -10,11 +10,11 @@ import {MarketplaceBaseStorage} from "./storage/MarketplaceBaseStorage.sol";
  * @title Base ERC20 implementation, excluding optional extensions
  */
 abstract contract MarketplaceBase is MarketplaceBaseInternal, IMarketplaceBase {
-    function fee() external view returns (uint16) {
+    function fee() external view returns (uint104) {
         return _fee();
     }
 
-    function mintFee() external view returns (uint16) {
+    function mintFee() external view returns (uint104) {
         return _mintFee();
     }
