+++ contracts/marketplace/base/interfaces/IMarketplaceBase.sol
@@ -10,9 +10,9 @@ import {MarketplaceBaseStorage} from "../storage/MarketplaceBaseStorage.sol";
  * @dev see https://github.com/ethereum/EIPs/issues/20
  */
 interface IMarketplaceBase is IMarketplaceBaseInternal {
-    function fee() external view returns (uint16);
+    function fee() external view returns (uint104);
 
-    function mintFee() external view returns (uint16);
+    function mintFee() external view returns (uint104);
 
     function decimals() external view returns (uint8);
